model/data/requirements.txt
flickrapi
numpy
pillow
tqdm
